+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- Example script to populate a dummy result DB
+
 ### Operations
 
 - `codeql.yml` and `condabuild.yml` run nightly instead of on every PR.